Create GettersTest
